+++ packages/toujou-overlay/tests/index.test.ts
@@ -1,7 +1,7 @@
 import { aTimeout, elementUpdated, expect, fixture, html } from '@open-wc/testing';
 
 import '../src/index';
-import { setOverlayCookie } from '../src/utils/cookie';
+import { setOverlayStorageValue } from '../src/utils/storage';
 import { ToujouOverlay } from '../src';
 import sinon from 'sinon';
 
@@ -10,6 +10,8 @@ describe('Toujou Overlay', () => {
 
   beforeEach(async () => {
     document.cookie.split(';').forEach(function(c) { document.cookie = c.replace(/^ +/, '').replace(/=.*/, '=;expires=' + new Date().toUTCString() + ';path=/'); });
+    localStorage.clear();
+    window.location.hash = '';
 
     element = await fixture(html`
       <toujou-overlay id="overlay" role="dialog" aria-modal="true">
@@ -45,7 +47,6 @@ describe('Toujou Overlay', () => {
     expect(element.querySelector('.overlay__text')).not.to.be.null;
     expect(element.querySelector('.overlay__buttons')).not.to.be.null;
     expect(element).to.have.attribute('opened');
-
   });
 
   it('it will close overlay on yes click', async () => {
@@ -82,8 +83,8 @@ describe('Toujou Overlay', () => {
     expect(element.querySelector('.overlay__buttons')).not.to.be.displayed;
   });
 
-  it('it will hide overlay if cookie is set ', async () => {
-    setOverlayCookie('overlay', 'accepted');
+  it('it will hide overlay if cookie is set', async () => {
+    setOverlayStorageValue('overlay', 'accepted');
     element = await fixture(html`
       <toujou-overlay id="overlay">
         <template><!--
@@ -124,7 +125,7 @@ describe('Toujou Overlay', () => {
     expect(element.querySelector('.overlay__text')).not.to.be.null;
   });
 
-  it('it will hide overlay if #aas is  ', async () => {
+  it('it will hide overlay if #aas is set via url', async () => {
     window.location.hash = '#aaa';
     element = await fixture(html`
       <toujou-overlay id="overlay">
@@ -206,5 +207,63 @@ describe('Toujou Overlay', () => {
     await elementUpdated(element);
 
     expect(document.body.classList.contains('toujou-overlay-open')).to.be.false;
+  });
+
+  it('it will set storage-mode by default on "session"', async () => {
+    await elementUpdated(element);
+    expect(element.getAttribute('storage-mode')).to.equal('session');
+  });
+
+  it('it will hide overlay if localStorage overlay value is set to "accepted" on storage mode "localStorage"', async () => {
+    localStorage.setItem('toujou-overlay-local', 'accepted');
+    element = await fixture(html`
+      <toujou-overlay id="local" storage-mode="localStorage">
+        <template><!--
+         <div class="overlay__text">
+            Hello World
+         </div>
+         <span class="overlay__warning" style="display:none;">Warning</span>
+         <div class="overlay__buttons">
+            <a class="overlay__button" data-overlay-value="yes">Yes</a>
+            <a class="overlay__button" data-overlay-value="no">No</a>
+         </div>
+
+        --></template>
+      </toujou-overlay>
+    `);
+    await elementUpdated(element);
+
+    expect(element.getAttribute('storage-mode')).to.equal('localStorage');
+    expect(element.querySelector('.overlay__text')).to.be.null;
+    expect(element.querySelector('.overlay__buttons')).to.be.null;
+    expect(element).not.to.have.attribute('opened');
+  });
+
+  it('it will set localStorage overlay value on close on storageMode localStorage', async () => {
+    element = await fixture(html`
+      <toujou-overlay id="local" storage-mode="localStorage">
+        <template><!--
+         <div class="overlay__text">
+            Hello World
+         </div>
+         <span class="overlay__warning" style="display:none;">Warning</span>
+         <button
+            class="overlay__close-button overlay-button-close-trigger"
+            aria-label="close"
+          ></button>
+
+        --></template>
+      </toujou-overlay>
+    `);
+
+    await elementUpdated(element);
+
+    const closeButton = element.querySelector('.overlay__close-button') as HTMLElement;
+    expect(closeButton).to.not.be.null;
+
+    closeButton.click();
+    await elementUpdated(element);
+
+    expect(localStorage.getItem('toujou-overlay-local')).to.be.equal('accepted');
   });
 });